add the suggested changes by @SimonDingemans, now the chunk tests are spread out, test_document_parser uses fixtures
