+++ src/database_builder_libs/utility/chunk/n_points_section.py
@@ -4,10 +4,10 @@ from dataclasses import dataclass
 from typing import Sequence
  
 from database_builder_libs.models.chunk import Chunk
-from database_builder_libs.models.chunk_strategy import ChunkingStrategy, RawSection
+from database_builder_libs.models.chunk_strategy import AbstractChunkingStrategy, RawSection
 
 @dataclass(slots=True)
-class SectionChunkingStrategy(ChunkingStrategy):
+class SectionChunkingStrategy(AbstractChunkingStrategy):
     """
     Produces exactly one ``Chunk`` per non-empty document section.
