+++ src/database_builder_libs/utility/chunk/summary_n_points.py
@@ -4,10 +4,10 @@ from dataclasses import dataclass
 from typing import Sequence
  
 from database_builder_libs.models.chunk import Chunk
-from database_builder_libs.models.chunk_strategy import ChunkingStrategy, RawSection
+from database_builder_libs.models.chunk_strategy import AbstractChunkingStrategy, RawSection
 
 @dataclass(slots=True)
-class SummaryAndNSectionsStrategy(ChunkingStrategy):
+class SummaryAndNSectionsStrategy(AbstractChunkingStrategy):
     """
     Produces one summary chunk followed by exactly ``n_sections`` body chunks.
